Implement server side validation

+++ src/actions/actions.ts
@@ -3,32 +3,55 @@
 import { prisma } from "@/lib/db";
 import { sleep } from "@/lib/utils";
 import { revalidatePath } from "next/cache";
-import { PetEssentials, PetId } from "@/lib/types";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
+import { Pet } from "@prisma/client";
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
 
+  const validatedPet = petFormSchema.safeParse(pet);
+  if (!validatedPet.success) {
+    return {
+      message: "Invalid pet data.",
+      success: false,
+    } as const;
+  }
+
+  let newPet: Pet;
   try {
-    await prisma.pet.create({
-      data: pet,
+    newPet = await prisma.pet.create({
+      data: validatedPet.data,
     });
   } catch (error) {
     console.error(error);
-    return { message: "Could not add pet" };
+    return { message: "Could not add pet", success: false } as const;
   }
 
   revalidatePath("/app", "layout");
+
+  return {
+    success: true,
+    pet: newPet,
+  } as const;
 }
 
-export async function editPet(petId: PetId, newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+  if (!validatedPetId.success || !validatedPet.success) {
+    return {
+      message: "Invalid pet data.",
+    };
+  }
 
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     console.error(error);
@@ -38,13 +61,20 @@ export async function editPet(petId: PetId, newPetData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: PetId) {
-  await sleep(3000);
+export async function deletePet(petId: unknown) {
+  await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data.",
+    };
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
